fix: delete executeCGIProcess file and move to Connection Class file
